Data is now loaded from ~/nevis-data, or from a path set in the environment variable NEVIS_PATH. The _data module has been refactored into three seperate modules.
