+++ nevis/__init__.py
@@ -2,29 +2,69 @@
 """
 Main module for Where's Ben Nevis (it's in Scotland).
 """
-import sys
 
-# Check Python version
-if sys.hexversion < 0x03060000:
-    raise RuntimeError('BenNevisServer requires Python 3.6 or newer.')
-
+#
 # Version number
-__version_tuple__ = (0, 0, 2)
+#
+__version_tuple__ = (0, 0, 3)
 __version__ = '.'.join(str(x) for x in __version_tuple__)
 
+#
+# Check Python version
+#
+import sys
+if sys.hexversion < 0x03060000:
+    raise RuntimeError('nevis requires Python 3.6 or newer.')
+del(sys)    # Don't expose as part of API
+
+#
+# Installed project root
+#
+import os, inspect  # noqa
+try:
+    frame = inspect.currentframe()
+    _DIR_MODULE = os.path.abspath(os.path.dirname(inspect.getfile(frame)))
+    _DIR_MODULE_DATA = os.path.join(_DIR_MODULE, '_bin')
+finally:
+    # Always manually delete frame
+    # https://docs.python.org/2/library/inspect.html#the-interpreter-stack
+    del(frame)
+del(os, inspect)    # Don't expose as part of API
+
+#
+# Terrain data
+#
+import os   # noqa
+
+_ENV_DATA = 'NEVIS_PATH'
+_DIR_DATA = os.environ.get(_ENV_DATA)
+if _DIR_DATA is None:
+    _DIR_DATA = os.path.join('~', 'nevis-data')
+_DIR_DATA = os.path.abspath(os.path.expanduser(_DIR_DATA))
+del(os)
+
+
+#
 # Create public API
-from ._data import (    # noqa
+#
+from ._bng import (    # noqa
     ben,
     Coords,
     dimensions,
     fen,
-    gb,
     Hill,
-    linear_interpolant,
     pub,
-    spacing,
-    spline,
     squares,
+)
+from ._os50 import (    # noqa
+    DataNotFoundError,
+    download_os50,
+    gb,
+    spacing,
+)
+from ._interpolation import (   # noqa
+    linear_interpolant,
+    spline,
 )
 from ._plot import (    # noqa
     plot,
@@ -37,7 +77,10 @@ from ._util import (    # noqa
 )
 
 
-def howdy(name='Server'):
+#
+# Version-related methods
+#
+def howdy(name='Local'):
     """ Say hi the old fashioned way. """
     print('')
     print('                |>          ')
@@ -49,6 +92,3 @@ def howdy(name='Server'):
     print('     /\/  \  /  \_/    \    ')
     print('    /      \/           \   ')
 
-
-# Don't exposure imported modules as part of the API
-del(sys)
